requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-text-splitters
langchain-classic
sentence-transformers==2.7.0
faiss-cpu
pypdf
python-dotenv